application-currency converter-services/requirements.txt
grpcio==1.39.0
grpcio-tools==1.39.0
requests==2.25.1
urllib3==1.26.5
